+++ apps/extension/cypress/stubs/baseDiamond.tsx
@@ -1,2 +1,0 @@
-const baseDiamond = () => null;
-export default baseDiamond;